+++ pkg/providers/gtm/data_akamai_gtm_cidrmap.go
@@ -18,7 +18,6 @@ type cidrMapDataSource struct {
 }
 
 type cidrMapDataSourceModel struct {
-	ID                types.String        `tfsdk:"id"`
 	Domain            types.String        `tfsdk:"domain"`
 	Name              types.String        `tfsdk:"map_name"`
 	DefaultDatacenter *defaultDatacenter  `tfsdk:"default_datacenter"`
@@ -110,11 +109,6 @@ func (d *cidrMapDataSource) Schema(_ context.Context, _ datasource.SchemaRequest
 	resp.Schema = schema.Schema{
 		MarkdownDescription: "GTM CIDR map data source.",
 		Attributes: map[string]schema.Attribute{
-			"id": schema.StringAttribute{
-				MarkdownDescription: "Identifier of the data source.",
-				DeprecationMessage:  "Required by the terraform plugin testing framework, always set to `gtm_cidrmap`.",
-				Computed:            true,
-			},
 			"domain": schema.StringAttribute{
 				Required:    true,
 				Description: "GTM domain name.",
@@ -165,7 +159,6 @@ func (m *cidrMapDataSourceModel) setAttributes(ctx context.Context, cidrMap *gtm
 	if diags.HasError() {
 		return diags
 	}
-	m.ID = types.StringValue("gtm_cidrmap")
 
 	return nil
 }